new base class framework for CTF
