+++ src/cryojax/simulator/_transfer_theory/transfer_theory.py
@@ -1,102 +1,12 @@
 from typing import Optional
-from typing_extensions import override
 
 import equinox as eqx
 import jax.numpy as jnp
-from jaxtyping import Array, Complex, Float
+from jaxtyping import Array, Complex
 
 from ...image.operators import FourierOperatorLike
-from ...internal import error_if_negative, error_if_not_fractional
 from .._instrument_config import InstrumentConfig
-from .base_transfer_theory import AbstractTransferFunction
-from .common_functions import compute_phase_shift_from_amplitude_contrast_ratio
-
-
-class ContrastTransferFunction(AbstractTransferFunction, strict=True):
-    """Compute an astigmatic Contrast Transfer Function (CTF) with a
-    spherical aberration correction and amplitude contrast ratio.
-
-    !!! info
-        `cryojax` uses a convention different from CTFFIND for
-        astigmatism parameters. CTFFIND returns defocus major and minor
-        axes, called "defocus1" and "defocus2". In order to convert
-        from CTFFIND to `cryojax`,
-
-        ```python
-        defocus1, defocus2 = ... # Read from CTFFIND
-        ctf = ContrastTransferFunction(
-            defocus_in_angstroms=(defocus1+defocus2)/2,
-            astigmatism_in_angstroms=defocus1-defocus2,
-            ...
-        )
-        ```
-    """
-
-    defocus_in_angstroms: Float[Array, ""]
-    astigmatism_in_angstroms: Float[Array, ""]
-    astigmatism_angle: Float[Array, ""]
-    spherical_aberration_in_mm: Float[Array, ""]
-    amplitude_contrast_ratio: Float[Array, ""]
-    phase_shift: Float[Array, ""]
-
-    def __init__(
-        self,
-        defocus_in_angstroms: float | Float[Array, ""] = 10000.0,
-        astigmatism_in_angstroms: float | Float[Array, ""] = 0.0,
-        astigmatism_angle: float | Float[Array, ""] = 0.0,
-        spherical_aberration_in_mm: float | Float[Array, ""] = 2.7,
-        amplitude_contrast_ratio: float | Float[Array, ""] = 0.1,
-        phase_shift: float | Float[Array, ""] = 0.0,
-    ):
-        """**Arguments:**
-
-        - `defocus_in_angstroms`: The mean defocus in Angstroms.
-        - `astigmatism_in_angstroms`: The amount of astigmatism in Angstroms.
-        - `astigmatism_angle`: The defocus angle.
-        - `spherical_aberration_in_mm`: The spherical aberration coefficient in mm.
-        - `amplitude_contrast_ratio`: The amplitude contrast ratio.
-        - `phase_shift`: The additional phase shift.
-        """
-        self.defocus_in_angstroms = jnp.asarray(defocus_in_angstroms)
-        self.astigmatism_in_angstroms = jnp.asarray(astigmatism_in_angstroms)
-        self.astigmatism_angle = jnp.asarray(astigmatism_angle)
-        self.spherical_aberration_in_mm = error_if_negative(spherical_aberration_in_mm)
-        self.amplitude_contrast_ratio = error_if_not_fractional(amplitude_contrast_ratio)
-        self.phase_shift = jnp.asarray(phase_shift)
-
-    @override
-    def __call__(
-        self,
-        frequency_grid_in_angstroms: Float[Array, "y_dim x_dim 2"],
-        voltage_in_kilovolts: Float[Array, ""] | float,
-    ) -> Float[Array, "y_dim x_dim"]:
-        """Compute the CTF as a JAX array.
-
-        **Arguments:**
-
-        - `frequency_grid_in_angstroms`:
-            The grid of frequencies in units of inverse angstroms. This can
-            be computed with [`cryojax.coordinates.make_frequency_grid`](https://mjo22.github.io/cryojax/api/coordinates/making_coordinates/#cryojax.coordinates.make_frequency_grid)
-        - `voltage_in_kilovolts`:
-            The accelerating voltage of the microscope in kilovolts. This
-            is converted to the wavelength of incident electrons using
-            the function [`cryojax.constants.convert_keV_to_angstroms`](https://mjo22.github.io/cryojax/api/constants/units/#cryojax.constants.convert_keV_to_angstroms)
-        """  # noqa: E501
-        # Convert degrees to radians
-        aberration_phase_shifts = self.compute_aberration_phase_shifts(
-            frequency_grid_in_angstroms, voltage_in_kilovolts=voltage_in_kilovolts
-        )
-        # Additional phase shifts
-        phase_shift = jnp.deg2rad(self.phase_shift)
-        amplitude_contrast_phase_shift = (
-            compute_phase_shift_from_amplitude_contrast_ratio(
-                self.amplitude_contrast_ratio
-            )
-        )
-        # Compute the CTF
-        return jnp.sin(
-            aberration_phase_shifts - (phase_shift + amplitude_contrast_phase_shift)
-        )
+from .transfer_function import ContrastTransferFunction
 
 
 class ContrastTransferTheory(eqx.Module, strict=True):
@@ -162,6 +72,7 @@ class ContrastTransferTheory(eqx.Module, strict=True):
             ctf_array = self.ctf(
                 frequency_grid,
                 voltage_in_kilovolts=instrument_config.voltage_in_kilovolts,
+                is_weak_phase_approximation=True,
             )
             # ... compute the contrast as the CTF multiplied by the exit plane
             # phase shifts
@@ -185,6 +96,49 @@ class ContrastTransferTheory(eqx.Module, strict=True):
             contrast_spectrum_at_detector_plane *= self.envelope(frequency_grid)
 
         return contrast_spectrum_at_detector_plane
+
+
+class WaveTransferTheory(eqx.Module, strict=True):
+    """An optics model that propagates the exit wave to the detector plane."""
+
+    ctf: ContrastTransferFunction
+
+    def __init__(
+        self,
+        ctf: ContrastTransferFunction,
+    ):
+        """**Arguments:**
+
+        - `ctf`: The contrast transfer function model.
+        """
+
+        self.ctf = ctf
+
+    def propagate_wavefunction_to_detector_plane(
+        self,
+        wavefunction_spectrum_at_exit_plane: Complex[
+            Array,
+            "{instrument_config.padded_y_dim} {instrument_config.padded_x_dim}",
+        ],
+        instrument_config: InstrumentConfig,
+    ) -> Complex[
+        Array, "{instrument_config.padded_y_dim} {instrument_config.padded_x_dim}"
+    ]:
+        """Apply the wave transfer function to the wavefunction in the exit plane."""
+        frequency_grid = instrument_config.padded_full_frequency_grid_in_angstroms
+        # Compute the wave transfer function
+        ctf_array = self.ctf(
+            frequency_grid,
+            voltage_in_kilovolts=instrument_config.voltage_in_kilovolts,
+            is_weak_phase_approximation=False,
+        )
+        # ... compute the contrast as the CTF multiplied by the exit plane
+        # phase shifts
+        wavefunction_spectrum_at_detector_plane = (
+            ctf_array * wavefunction_spectrum_at_exit_plane
+        )
+
+        return wavefunction_spectrum_at_detector_plane
 
 
 def _compute_contrast_from_ewald_sphere(
